requirements.txt
# ============================================================================
# Writing Hub — Requirements
# Django 5.2 LTS (ADR-083)
# ============================================================================

# Core
Django==5.2.6
gunicorn==23.0.0
whitenoise==6.11.0
python-decouple==3.8

# Django Extensions
django-filter==24.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
djangorestframework==3.16.1

# Error Tracking
sentry-sdk==2.18.0

# HTTP
requests==2.32.5
httpx>=0.27,<1
aiohttp==3.13.2

# Data
pydantic>=2.6,<3
pydantic-settings==2.12.0
structlog==25.4.0
asgiref==3.8.1

# ============================================================================
# iil Platform Packages
# ============================================================================

# LLM routing — DB-driven, sync_completion() / completion()
iil-aifw>=0.6.0,<1

# Prompt templates — PromptStack, 5-layer Jinja2, render_to_messages()
iil-promptfw>=0.5.1,<1

# Domain schemas — StyleProfile, CharacterProfile, WorldContext
iil-authoringfw>=0.6.0,<1

# WeltenHub REST client — WeltenClient, worlds/characters API
iil-weltenfw[django]>=0.1.0,<1

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.10
dj-database-url>=2.0.0

# ============================================================================
# Async Tasks
# ============================================================================
celery==5.5.3
redis==7.0.1
django-celery-beat==2.8.1

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.9.0.post0
PyYAML>=6.0
python-dotenv==1.0.0
markdown==3.4.1
jsonschema==4.25.1
orjson==3.11.4
tenacity>=8.2

# Document Processing
python-docx==1.2.0
PyPDF2==3.0.1
pdfplumber==0.11.7

# Image Processing
Pillow==12.0.0

# Security
cryptography==46.0.3
PyJWT==2.10.1

# LLM provider (transitive via aifw/litellm)
litellm>=1.30
